Adds prefix to libtomcrypt functions
Adds a prefix to all libtomcrypt functions to avoid naming conflicts with other libraries.
This change ensures that the libtomcrypt functions are uniquely identified, preventing potential collisions when the library is used in conjunction with other libraries that may have functions with the same names.

The prefix is defined in the Package.swift file and is prepended to all libtomcrypt function names using a macro.
This ensures that the correct functions are called, even if there are other functions with the same names in the program.